requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
python-multipart==0.0.12
aiohttp>=3.9.0
requests>=2.32.0
